Retain & Filter refactor (#46)
`sliceext.Retain` & `sliceext.Filter` to not shuffle data in the
underlying slice array but create new slice referencing the data
instead. In practice, it can cause unexpected behaviour and users
expectations not met when the same data is also referenced elsewhere. If
anyone still requires a `shuffle` implementation for efficiency I'd be
happy to add a separate function for that as well.